pyproject.toml
[tool.poetry]
name = "play-hackathon-beszketnyky"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.0"
uvicorn = {version = "0.32.0", extras = ["standard"]}
requests = "^2.32.5"
beautifulsoup4 = "4.12.3"
playwright = "^1.55.0"
pandas = "^2.3.3"
httpx = "0.27.2"
python-dotenv = "^1.1.1"
langchain = "^1.0.2"
openai = "1.55.3"
langgraph = "^1.0.1"
pydantic = "2.9.2"
trafilatura = "^1.6.6"
python-multipart = "0.0.12"
sqlalchemy = "2.0.36"
psycopg2-binary = "2.9.9"
asyncpg = "0.29.0"
redis = "5.0.8"
pydantic-settings = "2.6.1"
lxml = "5.3.0"
aiosqlite = "0.20.0"
boto3 = "^1.40.59"
aiosmtplib = "^5.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
black = "^25.9.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
